Tests/LibWeb/Text/expected/hit_testing/caret-and-element-from-point-relpos-wrapped-span.txt

rects: 2
elementFromPoint is span: true
caret node is span text: true
caret offset: 7
